+++ sample/counter/src/androidMain/kotlin/io/github/pavelannin/keemun/sample/counter/CounterFeatureScope.kt
@@ -3,7 +3,7 @@ package io.github.pavelannin.keemun.sample.counter
 import androidx.compose.runtime.Composable
 import com.arkivanov.decompose.ComponentContext
 import io.github.pavelannin.keemun.core.store.Store
-import io.github.pavelannin.keemun.decompose.KeemunComponentFeatureConnector
+import io.github.pavelannin.keemun.decompose.KeemunComponentConnector
 import io.github.pavelannin.keemun.sample.counter.features.counter.Counter
 import io.github.pavelannin.keemun.sample.counter.features.counter.CounterViewState
 import io.github.pavelannin.keemun.sample.counter.features.counter.ExternalMsg
@@ -11,7 +11,7 @@ import io.github.pavelannin.keemun.sample.counter.features.counter.counterFeatur
 
 actual class CounterFeatureScope {
 
-    fun counter(componentContext: ComponentContext) = KeemunComponentFeatureConnector(
+    fun counter(componentContext: ComponentContext) = KeemunComponentConnector(
         componentContext = componentContext,
         featureParams = counterFeatureParams()
     )