+++ sample/counter/src/iosMain/kotlin/io/github/pavelannin/keemun/sample/counter/CounterFeatureScope.kt
@@ -1,8 +1,8 @@
 package io.github.pavelannin.keemun.sample.counter
 
 import io.github.pavelannin.keemun.sample.counter.features.counter.counterFeatureParams
-import io.github.pavelannin.keemun.swiftui.KeemunNativeFeatureConnector
+import io.github.pavelannin.keemun.swiftui.KeemunNativeConnector
 
 actual class CounterFeatureScope {
-    fun counter() = KeemunNativeFeatureConnector(featureParams = counterFeatureParams())
+    fun counter() = KeemunNativeConnector(featureParams = counterFeatureParams())
 }