+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# v1.1.0
+## Core
+- Renamed the `init` parameter to `start` from `StoreParams`
+- Removed `PreInitEffect` from `StoreParams`
+- Renamed `previous` to `savedState` from `Start`
+
+## Decompose
+- Renamed `KeemunComponentFeatureConnector` to `KeemunComponentConnector`
+
+## SwiftUI
+- Renamed `KeemunNativeFeatureConnector` to `KeemunNativeConnector`
+
+
 # v1.0.1
 - Optimization of swiftui connector dependencies